fix new branch

+++ backend/microservice2/src/main/java/tn/esprit/microservice2/Model/Course.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,7 +23,7 @@ public class Course {
     @Column(length = 1000)
     private String description;
 
-    private BigDecimal price;
+    private double price;
 
     @Column(nullable = false)
     private Integer durationInMonths;
@@ -56,11 +55,11 @@ public class Course {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
